// packages/adapter-codex/src/generated/v2/ThreadQueueAddParams.ts
// GENERATED CODE! DO NOT MODIFY BY HAND!

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { UserInput } from "./UserInput.ts";

export type ThreadQueueAddParams = {
  threadId: string;
  input: Array<UserInput>;
  clientUserMessageId: string;
};